feat: refactor agent creation into separate file
